Add x-abac-policy claim to internal issued JWT
Having the ABAC policy part of the JWT used to communicate to the
application prevents bypassing ABAC policy by tampering with the
(unsigned) X-ABAC-Policy http header.

Tampering with the JWT is not possible because it is signed; omitting
the JWT is also not possible because the application requires a signed
JWT to access any resources at all.