Refresh session-based OAuth2 authentication tokens
When using a session to perform OAuth2 login instead of a bearer token; the ID token is not refreshed at all.
A new access token is only obtained when a token is explicitly requested from the OAuth2AutorizedClientManager,
so we should to that before issuing a substitution-token for session-based authentication.

We don't want to blindly issue a new substitution-token with an expiry beyond the access token, because
that would allow minting new access tokens that are still valid, even when the access token itself has expired,
or the user's session was revoked in the identity provider.

Access-token based authentication is not affected by this change, as the
access token itself is per-request and must already be valid anyways.