Split up `AuthenticationInformationResolver` into components
There are multiple ways in which the `AuthenticationInformation` can be resolved (via oidc, jwt or oauth2 access token)
Not all of the dependencies of resolvers are available at runtime, so select the resolvers that are,
and make them return `AuthenticationInformation` when known.

When using session-based authentication, the tokens described in the spring-security `Authentication` might already
have expired. In that case, we need a fresh token to verify that the users' session in the issuer is still
active by refreshing the authorization token.

Since some of the user claims may also be changed, and a websession remains
active as long as the user is active, user claims can be fetched on
demand by a JwtClaimsResolver. Or when they are very unlikely to change or not security-sensitive,
the stale claims can be reused from the spring security `Authentication`.