Make it possible to bypass OPA authorization for runtime platform endpoints
When performing a request to tokenmonger (/.contentgrid/authentication/** on an app domain), the gateway performs a query to OPA to authorize the request.

However, the ContentGrid application itself (which determines the OPA policy) does not know about the /.contentgrid/authentication endpoint (and it also does not handle that endpoint), so it does not have any rule permitting access.

It would be logical for the gateway to bypass authorization with OPA for endpoints that are configured on it:

1. Available gateway endpoints are dependent on the runtime platform setup.
   It is possible for some runtime platforms to not support certain endpoints, and it is also possible for more endpoints to be added later on (without having to regenerate and redeploy all applications)
   If an endpoint is not available, and the application defines access allowed in the policy, it will allow access to that URL in the ContentGrid app; which may be unexpected
2. ContentGrid applications have no knowledge about these endpoints being intercepted, so why would they have to set up a policy to handle that
3. The upstreams of gateway endpoints are in a better position to do authorization of the incoming requests. They can still query OPA themselves (and maybe with different parameters) to determine some authorization.