Fix oidc tests for keycloak 25
The user profile that is now enabled in keycloak 25 enforces that email, first and last name are set by the user.
So we need to set them, otherwise we encounter an interstitial after logging in and before the authorization code
is sent back to the gateway. That breaks the test because it doesn't know about this step.

Simply providing the necessary details beforehand when the user is created solves that issue.