Do not cache client registration indefinitely
This is especially a problem if the client registration fails to resolve
due to a transient failure, where the failure gets cached forever.

However, it may also be an issue that a successful response is cached
forever, as the response from the openid-configuration endpoint can
change.

To avoid flooding an oidc server with requests, cache successful
responses for 10 minutes and failures for 1 minute.